% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdd}
\alias{mdd}
\title{Calculate Maximum Drawdown (MDD)}
\usage{
mdd(returns)
}
\arguments{
\item{returns}{A numeric vector of returns. Each value represents the return
for a specific period.}
}
\value{
A numeric value representing the maximum drawdown.
}
\description{
This function calculates the maximum drawdown (MDD) for a series of returns.
The MDD represents the largest percentage drop from a peak to a trough during
the investment period.
}
\examples{
returns <- c(0.01, 0.02, 0.015, -0.02, 0.03)
mdd(returns)
}
